[ABCVm] fix pushshort opcode for negative values

+++ src/scripting/abc_interpreter.cpp
@@ -4353,7 +4353,7 @@ void ABCVm::preloadFunction(SyntheticFunction* function)
 				int32_t p = code.tellg();
 				state.preloadedcode.push_back((uint32_t)opcode);
 				state.oldnewpositions[code.tellg()] = (int32_t)state.preloadedcode.size();
-				int32_t value = code.readu32();
+				int32_t value = (int32_t)(int16_t)code.readu32();
 				uint16_t index = value;
 				state.preloadedcode.back().pcode.arg3_int=value;
 				if (state.jumptargets.find(p) != state.jumptargets.end())
